requirements.txt
click
PyYAML
requests
tiktoken
sentence-transformers
scikit-learn
scipy
numpy
GitPython
PyGithub
watchdog
tree-sitter